+++ DTCD-WorkspacePanel/src/Plugin.js
@@ -7,10 +7,13 @@ import {
   EventSystemAdapter,
   WorkspaceSystemAdapter,
   RouteSystemAdapter,
+  LogSystemAdapter,
 } from './../../DTCD-SDK/index';
 
 export class WorkspacePanel extends AppPanelPlugin {
   #vueComponent;
+  #logSystem;
+  #eventSystem;
 
   #config = {
     elementSize: 'medium',
@@ -24,13 +27,26 @@ export class WorkspacePanel extends AppPanelPlugin {
     super();
 
     const eventSystem = new EventSystemAdapter('0.4.0', guid);
+    this.#eventSystem = eventSystem;
+
+    this.#eventSystem.registerPluginInstance(this, ['WorkspaceDeleted']);
+
     const interactionSystem = new InteractionSystemAdapter('0.4.0');
     const workspaceSystem = new WorkspaceSystemAdapter('0.4.0');
     const router = new RouteSystemAdapter('0.1.0');
+    this.#logSystem = new LogSystemAdapter('0.5.0', guid, pluginMeta.name);
 
     const { default: VueJS } = this.getDependence('Vue');
 
-    const data = { guid, interactionSystem, eventSystem, workspaceSystem, plugin: this, router };
+    const data = {
+      guid,
+      interactionSystem,
+      eventSystem,
+      workspaceSystem,
+      plugin: this,
+      router,
+      logSystem: this.#logSystem,
+    };
 
     const panel = new VueJS({
       data: () => data,
@@ -61,6 +77,7 @@ export class WorkspacePanel extends AppPanelPlugin {
 
   getFormSettings() {
     const { selectedElement } = this.#vueComponent;
+    const { permissions } = selectedElement;
 
     const fields = [
       {
@@ -77,8 +94,7 @@ export class WorkspacePanel extends AppPanelPlugin {
         ],
       },
       {
-        component: 'title',
-        innerText: 'Настройки элемента',
+        component: 'divider',
       },
     ];
 
@@ -88,6 +104,10 @@ export class WorkspacePanel extends AppPanelPlugin {
         innerText: 'Выберете один из элементов рабочего стола',
       });
     } else {
+      const disabledAttrEditBtn = permissions.update ? {} : { disabled: true };
+      const disabledAttrDelBtn = permissions.delete ? {} : { disabled: true };
+      const disabledAttrExpoBtn = permissions.read ? {} : { disabled: true };
+
       fields.push(
         ...[
           {
@@ -95,43 +115,48 @@ export class WorkspacePanel extends AppPanelPlugin {
             innerText: `Выбран: ${selectedElement.title}`,
           },
           {
-            component: 'subtitle',
-            innerText: 'Удалить элемент',
-          },
-          {
             component: 'button',
-            innerText: 'Изменить',
+            innerText: 'Изменить элемент',
             panelRow: 'row2',
             column: '50',
             attrs: {
               theme: 'theme_blueSec',
               width: 'full',
               size: 'small',
+              ...disabledAttrEditBtn,
             },
             handler: {
               event: 'click',
-              callback: () => {
-                this.#vueComponent.editElement(selectedElement);
-              },
+              callback: permissions.update
+                ? () => {
+                  this.#vueComponent.editElement(selectedElement);
+                }
+                : undefined,
             },
           },
           {
             component: 'button',
-            innerText: 'Удалить',
+            innerText: 'Удалить элемент',
             panelRow: 'row2',
             column: '50',
             attrs: {
               theme: 'theme_red',
               width: 'full',
               size: 'small',
+              ...disabledAttrDelBtn,
             },
             handler: {
               event: 'click',
-              callback: () => {
-                const elemType = selectedElement.is_dir ? 'папку' : 'дашборд';
-                const isDelete = confirm(`Удалить ${elemType} "${selectedElement.title}"?`);
-                isDelete && this.#vueComponent.deleteElement(selectedElement);
-              },
+              callback: permissions.delete
+                ? async () => {
+                  const elemType = selectedElement.is_dir ? 'папку' : 'дашборд';
+                  const isDelete = confirm(`Удалить ${elemType} "${selectedElement.title}"?`);
+                  if (isDelete) {
+                    await this.#vueComponent.deleteElement(selectedElement);
+                    this.#eventSystem.publishEvent('WorkspaceDeleted', {guid: this.guid});
+                  }
+                }
+                : undefined,
             },
           },
         ]
@@ -140,19 +165,20 @@ export class WorkspacePanel extends AppPanelPlugin {
       if (!selectedElement.is_dir) {
         fields.push({
           component: 'button',
-          innerText: 'Экспортировать',
-          panelRow: 'row3',
-          column: '100',
+          innerText: 'Экспортировать элемент',
           attrs: {
             theme: 'theme_blueSec',
             width: 'full',
             size: 'small',
+            ...disabledAttrExpoBtn,
           },
           handler: {
             event: 'click',
-            callback: () => {
-              this.#vueComponent.exportConfiguration(selectedElement);
-            },
+            callback: permissions.read
+              ? () => {
+                this.#vueComponent.exportConfiguration(selectedElement);
+              }
+              : undefined,
           },
         });
       }