+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.6.0]
+
+### Added
+
+- publish of event of workspace delete
+- logging of interactions with the server (deleting, creation, exporting, etc.)
+- permissions creating, deleting, editing and reading elements
+
+### Changed
+
+- form settings for element
+
+## Fixed
+
+- workspace element colors are not applied or applied incorrectly
+
 ## [0.5.0]
 
 ### Added